backend/pyproject.toml
[project]
name = "agent-harness"
version = "0.1.0"
description = "Deep Research → Publishable Content Pipeline with Multi-Agent Orchestration"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "googlesearch-python>=1.3.0",
    "pydantic>=2.9.0",
    "sse-starlette>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-socketio>=5.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
